mow we have log in & registration

+++ base/views.py
@@ -1,8 +1,17 @@
+#from email import message
+from django.http import HttpResponse
+
 from django.shortcuts import render, redirect
+
+from django.contrib import messages
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
+
 from django.db.models import Q
 
-
-from random import randint
+from random import choice
 
 from .models import Room, Topic
 from .forms import RoomForm
@@ -36,9 +45,15 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 def random_room(request):
+    
     rooms = Room.objects.all()
-    return redirect('room', randint(1,len(rooms)))
+    title = choice(rooms)
+    
+    roomid = Room.objects.get(name=title).id
+    
+    return redirect('room', roomid)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == "POST":
@@ -49,11 +64,13 @@ def create_room(request):
     context={'form':form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
-
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('<h1>Go away!</h1>')
 
     if request.method == 'POST':
         form=RoomForm(request.POST, instance=room)
@@ -64,10 +81,60 @@ def update_room(request, pk):
     context ={'form':form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url='login')
 def delete_room (request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('<h1>Go away!</h1>')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
     return render(request, 'base/delete.html', {'obg':room})
+
+
+def login_user(request):
+    page = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+    if request.method =="POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, "U're not existing! yet.. ")
+        
+        user = authenticate(request, username=username, password=password)
+
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+             messages.error(request, "Username or password does not exist ")
+    context = {'page':page}
+    return render(request, 'base/login_register.html', context)
+
+def logout_user(request):
+    logout(request)
+    return redirect('home')
+
+def register_user(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if request.method =="POST":
+        form= UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username =user.username.lower()
+            
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, " registration error ")
+
+    return render(request, 'base/login_register.html', {'form':form})
